ncdf {ncdf}	R Documentation

Read, write, and create netCDF files (v1.6)

Description

Read from or write to existing netCDF format files, or create new ones.

Details

The netCDF data file format from Unidata is a platform-independent, binary file that also contains metadata describing the contents and format of the data in the file. NetCDF files contain one or more variables, which are structured as regular N-dimensional arrays. They also contain dimensions, which describe the extent of the variables' arrays. Data can be read from or written to variables in arbitrary hyperslabs. The R package 'ncdf' allows reading from, writing to, and creation of netCDF files. Note that the netCDF library must already be installed on your machine for this R interface to the library to work.

If you are absolutely new to netCDF files, they can be a little overwhelming, so here is a brief sketch of what documentation you need to read next.

If you want to READ data from an already-existing netCDF file, first call open.ncdf to open the file, then call get.var.ncdf to read the data from the file.

If you want to WRITE data to a new netCDF file, first call dim.def.ncdf to define the dimensions that your data exists along (for example, perhaps latitude and longitude), then call var.def.ncdf to define a variable in the netCDF file that will hold your data, then call create.ncdf to create the netCDF file, then call put.var.ncdf to write your data to the newly created netCDF file.

This is version 1.6 of the ncdf library.

Author(s)

David W. Pierce dpierce@ucsd.edu

References

http://www.unidata.ucar.edu/packages/netcdf/

See Also

att.put.ncdf, att.get.ncdf, close.ncdf, create.ncdf, dim.def.ncdf, get.var.ncdf, put.var.ncdf, open.ncdf, print.ncdf, set.missval.ncdf, sync.ncdf, var.def.ncdf. redef.ncdf.

[Package ncdf version 1.6.8 Index]