close.ncdf {ncdf}	R Documentation

Close a netCDF File

Description

Closes an open netCDF file, which flushes any unwritten data to disk.

Usage

 close.ncdf( con, ... )

Arguments

con
An object of class ncdf (as returned by either function open.ncdf() or function create.ncdf(), indicating what file to read from.

...
Other arguments passed to or from other methods.

Details

Data in a netCDF file might be cached in memory, for better performance. This data is written out when the file is closed. Therefore, always remember to close the file when done with it.

Author(s)

David W. Pierce dpierce@ucsd.edu

References

http://www.unidata.ucar.edu/packages/netcdf/

See Also

sync.ncdf.

Examples

## Not run: nc <- open.ncdf("salinity.nc")
## Not run: data <- get.var.ncdf( nc )	# Read the "only" var in the file
## Not run: close.ncdf(nc)

[Package ncdf version 1.6.8 Index]