enddef.ncdf {ncdf}	R Documentation

Takes a netCDF file out of define mode

Description

Changes a netCDF that is currently in define mode back into data mode.

Usage

 enddef.ncdf( nc )

Arguments

nc
An object of class ncdf (as returned by either function open.ncdf() or function create.ncdf(), indicating what file to read from.

Details

NetCDF files can be in "define mode", at which time dimensions and variables can be defined, or new attributes added to a file, or in "data mode", at which time data can be read from the file. This call puts a file that is currently in define mode back into data mode.

Note

The typical user will never need this call, nor will ever have to worry about "define mode" or "data mode". THIS CALL IS PROVIDED FOR ADVANCED USERS ONLY! If the user goes through this package's standard functional interface, the file will always automatically be set to whatever mode it needs to be in without the user having to do anything. In particular, the call to write an attribute (att.put.ncdf) handles this automatically. An example of the kind of situation where you would need this call is if you are adding a new variable to an already-existing netCDF file. This case is not really handled by this package.

Author(s)

David W. Pierce dpierce@ucsd.edu

References

http://www.unidata.ucar.edu/packages/netcdf/

See Also

redef.ncdf.

Examples

# This function is for advanced useage only, and will never
# be needed by the typical users R code.

[Package ncdf version 1.6.8 Index]