print.ncdf {ncdf}	R Documentation

Print Information About a netCDF File

Description

Prints information about a netCDF file, including the variables and dimensions it contains.

Usage

 ## S3 method for class 'ncdf'
print( x, ... )

Arguments

x
An object of class "ncdf".

...
Extra arguments are passed to the generic print function.

Details

NetCDF files contain variables, which themselves have dimensions. This routine prints out useful information about a netCDF file's variables and dimensions. It is overloaded on the regular print function, so if "nc" is an object of class "ncdf", then just calling print(nc) will suffice. Objects of class "ncdf" are returned from open.ncdf.

Author(s)

David W. Pierce dpierce@ucsd.edu

References

http://www.unidata.ucar.edu/packages/netcdf/

See Also

var.def.ncdf

Examples

# Open a netCDF file, print information about it
nc <- open.ncdf( "salinity.nc" )
print(nc)

[Package ncdf version 1.6.8 Index]