version.ncdf {ncdf}	R Documentation

Report version of ncdf package

Description

Returns a string that is the version number of the ncdf package.

Usage

 version.ncdf()

Details

Note that the returned value it is a string, not a floating point number.

Value

A string (not float) that is the version number of the ncdf package.

Author(s)

David W. Pierce dpierce@ucsd.edu

References

http://www.unidata.ucar.edu/packages/netcdf/

[Package ncdf version 1.6.8 Index]